legacy-requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
aniso8601==9.0.1
anyio==3.6.1
appdirs==1.4.4
arrow==1.2.2
asgiref==3.5.2
attrs==19.3.0
backoff==1.11.1
black==22.6.0
boto3==1.11.9
botocore==1.14.9
CacheControl==0.12.6
cachetools==4.2.4
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==8.0.2
colorama==0.4.4
configobj==5.0.9
contextlib2==21.6.0
cryptography==2.8
cssselect==1.1.0
cssselect2==0.6.0
cssutils==1.0.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.4
distro==1.7.0
dj-database-url==0.5.0
Django==2.2.9
django-anymail==7.1.0
django-braces==1.14.0
django-composite-field==1.1.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-fsm==2.7.0
django-import-export==2.0.1
django-ltree==0.5.2
django-model-utils==4.0.0
django-pandas==0.6.4
django-sortedm2m==3.0.2
django-storages==1.8
django-taggit==1.2.0
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.5.0
docutils==0.15.2
drf-nested-routers==0.91
entrypoints==0.4
et-xmlfile==1.0.1
fastapi==0.75.2
filelock==3.5.1
# gevent==21.8.0
google-api-core==1.31.6
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-pubsub==2.8.0
google-cloud-run==0.1.0
google-cloud-storage==1.28.1
google-cloud-tasks==2.8.1
google-resumable-media==0.5.1
googleapis-common-protos==1.56.3
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.47.0
gunicorn==20.0.4
h11==0.13.0
html5lib==1.1
idna==2.8
importlib-metadata==4.8.1
isodate==0.6.0
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.9.4
jsonfield==3.1.0
jsonschema==4.6.1
libcst==0.4.6
lxml==4.6.2
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==1.0.3
mypy-extensions==0.4.3
nspektr==0.4.0
numpy==1.26.4
oauthlib==3.2.0
odfpy==1.4.0
openpyxl==3.0.3
ordered-set==4.0.2
packaging==20.1
pandas==1.3.5
pathspec==0.9.0
pep517==0.12.0
pikaur==1.10.1
Pillow==10.2.0
Pint==0.8.1
platformdirs==2.5.1
pluggy==0.13.1
ply==3.11
portpicker==1.5.2
premailer==3.7.0
progress==1.6
promise==2.3
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.8.4
py==1.8.1
pyalpm==0.10.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.19
pycryptodome==3.9.4
pydantic==1.9.0
PyGObject==3.42.1
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==2.4.6
pyphen==0.12.0
pyrsistent==0.18.1
pytest==5.3.5
pytest-django==3.8.0
pytest-order==1.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.1
resolvelib==0.5.5
retrying==1.3.3
rollbar==0.14.7
rsa==4.8
s3transfer==0.3.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.3.0
starlette==0.17.1
tablib==0.14.0
terminator==2.1.1
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
toposort==1.6
tornado==6.2
trove-classifiers==2022.5.19
typing-inspect==0.7.1
typing_extensions==4.3.0
urllib3==1.25.8
uvicorn==0.17.6
validate==5.1.0.dev0
validate-pyproject==0.9
vega-datasets==0.9.0
virtualenv==20.11.0
wcwidth==0.1.8
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
